requirements.txt
# Enhanced AI Game Testing Platform - Requirements
# Phase 1-2 Accuracy Improvements

# Core Dependencies (Existing)
streamlit>=1.28.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
pillow>=10.0.0

# New Dependencies for Phase 1-2

# Computer Vision & OCR
opencv-python>=4.8.0        # Template matching, image processing
pytesseract>=0.3.10         # OCR text extraction
numpy>=1.24.0               # Image array operations

# Optional but recommended
# easyocr>=1.7.0            # Alternative OCR (more accurate but slower)

# System Requirements:
# - Python 3.8+
# - ADB (Android Debug Bridge) installed and in PATH
# - Tesseract OCR installed for pytesseract
#   - Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
#   - Mac: brew install tesseract
#   - Linux: apt-get install tesseract-ocr

# Installation:
# pip install -r requirements_enhanced.txt

# Note: Make sure to install Tesseract OCR separately for your OS